Move attribute processing functions to kani_middle.
