+++ kani-compiler/src/kani_middle/mod.rs
@@ -2,4 +2,5 @@
 // SPDX-License-Identifier: Apache-2.0 OR MIT
 //! This module contains code that are backend agnostic. For example, MIR analysis
 //! and transformations.
+pub mod attributes;
 pub mod reachability;
